% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.pannet}
\alias{fitted.pannet}
\title{Fitted values for a pannet model}
\usage{
\method{fitted}{pannet}(object, ...)
}
\arguments{
\item{object}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\value{
The fitted response values on the response scale.
}
\description{
Fitted values for a pannet model
}
