% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.pannet}
\alias{plot.pannet}
\title{Plot training history for a pannet model}
\usage{
\method{plot}{pannet}(x, ...)
}
\arguments{
\item{x}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\description{
Plot training history for a pannet model
}
