% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.pannet}
\alias{print.pannet}
\title{Print a fitted pannet model}
\usage{
\method{print}{pannet}(x, ...)
}
\arguments{
\item{x}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\description{
Print a fitted pannet model
}
