% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.pannet}
\alias{residuals.pannet}
\title{Residuals from a fitted pannet model}
\usage{
\method{residuals}{pannet}(object, ...)
}
\arguments{
\item{object}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\value{
Residuals on the response scale.
}
\description{
Residuals from a fitted pannet model
}
