% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{time_effects}
\alias{time_effects}
\title{Retrieve learned time effects from a fitted pannet model}
\usage{
time_effects(object, ...)
}
\arguments{
\item{object}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\value{
Named numeric vector of learned time effects, or \code{NULL}.
}
\description{
Retrieve learned time effects from a fitted pannet model
}
