% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.pannet}
\alias{summary.pannet}
\title{Summarise a fitted pannet model}
\usage{
\method{summary}{pannet}(object, ...)
}
\arguments{
\item{object}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\value{
A list of class \code{"summary.pannet"}.
}
\description{
Summarise a fitted pannet model
}
