% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{training_history}
\alias{training_history}
\title{Extract the training history from a pannet model}
\usage{
training_history(object, ...)
}
\arguments{
\item{object}{A \code{"pannet"} object.}

\item{...}{Unused.}
}
\value{
A list containing loss curves and training metadata.
}
\description{
Extract the training history from a pannet model
}
